[Update] 'HomeController'를 'ViewController'로 변경
1. 'HomeController'를 'ViewController'로 변경
2. 뷰를 가져오는 메소드는 모두 여기로 모음
3. 카카오로그인 컨트롤러에 있는 로그인 뷰 가져오는 메소드 뷰컨트롤러로 이동
