--- src/main/java/com/sparta/lafesta/common/controller/HomeController.java
@@ -1,12 +1,0 @@
-package com.sparta.lafesta.common.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-    @GetMapping("/")
-    public String home() {
-        return "main";
-    }
-}
